refactor(insights): address review feedback on billing features
- Persist the data volume cap through the subresource API instead of writing to the
  filesystem from the resource provider, following the RedisServiceControlPlane pattern
  for a singleton subresource. InsightsService now declares the subresource.
- Move UpdateBillingFeaturesRequest to Models/Requests.
